\name{usbroadband-package}
\alias{usbroadband-package}
\alias{usbroadband}
\docType{package}
\title{Data for the U.S. National Broadband Map}
\description{Access to the data underlying the National Broadband Map (\url{http://www.broadbandmap.gov/}).}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{\url{http://www.broadbandmap.gov/developer/}}
\keyword{ package }
%\seealso{}
%\examples{}
